requirements.txt
Django
Pillow
django-jazzmin
django-cors-headers
django-extensions
django-filter
djangorestframework
djangorestframework-simplejwt
drf-spectacular
drf-spectacular-sidecar
PyJWT